Discovered bug, implemented temp workaround
Discovered that the file path string for the file browser system can’t
exceed 78 characters for some reason.

Temp solution: basic error messaging to warns user in-game when file
names are too long (and proceeds to not load anything, waiting for the
user to shorten the name).

Will try to discover source of bug and see if I can extend limit, but
this is low priority.